+++ e2e/a11y.spec.ts
@@ -9,6 +9,7 @@ const routes = [
   '/play',
   '/play/scenario/vote-sanrakshan',
   '/map',
+  '/google-services',
   '/pwd',
   '/yatra',
 ];